StarterService renamed to new-pwa-starter
