+++ src/extension.ts
@@ -1,7 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import { setUpLocalPwaStarterRepository } from './services/StarterService';
+import { setUpLocalPwaStarterRepository } from './services/new-pwa-starter';
 import { handleServiceWorkerCommand } from './services/service-worker';
 
 // this method is called when your extension is activated
